merge

+++ src/usr/bin/opentelemetry_shell_api.sh
@@ -7,18 +7,23 @@
 
 _otel_remote_sdk_pipe=$(\mktemp -u)_opentelemetry_shell_$$.pipe
 _otel_shell=$(\readlink /proc/$$/exe | \rev | \cut -d/ -f1 | \rev)
-_otel_commandline_override="$OTEL_SHELL_COMMANDLINE_OVERRIDE"
+if \[ "$OTEL_SHELL_COMMANDLINE_OVERRIDE_SIGNATURE" = "$PPID" ] || \[ "$OTEL_SHELL_COMMANDLINE_OVERRIDE_SIGNATURE" = "0" ]; then _otel_commandline_override="$OTEL_SHELL_COMMANDLINE_OVERRIDE"; fi
 unset OTEL_SHELL_COMMANDLINE_OVERRIDE
+unset OTEL_SHELL_COMMANDLINE_OVERRIDE_SIGNATURE
 unset OTEL_SHELL_SPAN_NAME_OVERRIDE
 unset OTEL_SHELL_SPAN_KIND_OVERRIDE
 unset OTEL_SHELL_SPAN_ATTRIBUTES_OVERRIDE
 unset OTEL_SHELL_SUPPRESS_LOG_COLLECTION
 
+_otel_command_real_self() {
+  \ps -p $$ -o args | \grep -v COMMAND || \cat /proc/$$/cmdline | \tr -d '\000'
+}
+
 _otel_command_self() {
   if \[ -n "$_otel_commandline_override" ]; then
-    \echo $_otel_commandline_override
+    \echo "$_otel_commandline_override"
   else
-    \ps -p $$ -o args | \grep -v COMMAND || \cat /proc/$$/cmdline
+    _otel_command_real_self
   fi
 }
 
@@ -113,7 +118,7 @@ otel_span_error() {
 otel_span_attribute() {
   local span_id=$1
   shift
-  local kvp="$@"
+  local kvp="$*"
   _otel_sdk_communicate "SPAN_ATTRIBUTE $span_id $kvp"
 }
 
@@ -175,7 +180,7 @@ _otel_escape_arg() {
     local do_escape=1
   else
     case "$1X" in
-      *[[:space:]\&\<\>\|\'\"\(\)\`!\$\;]*) local do_escape=1 ;;
+      *[[:space:]\&\<\>\|\'\"\(\)\`!\$\;\\]*) local do_escape=1 ;;
       *) local do_escape=0 ;;
     esac
   fi
@@ -206,8 +211,11 @@ _otel_line_join() {
 _otel_call() {
   # old versions of dash dont set env vars properly
   # more specifically they do not make variables that are set in front of commands part of the child process env vars but only of the local execution environment
-  \eval "$( { \printenv; \set; } | \grep '^OTEL_' | \cut -d= -f1 | \sort -u | \awk '{ print $1 "=\"$" $1 "\"" }' | _otel_line_join)" "\\$(_otel_escape_args "$@")"
-  # if \[ "$?" -ne 0 ]; then \echo "$( { \printenv; \set; } | \grep '^OTEL_' | \cut -d= -f1 | \sort -u | \awk '{ print $1 "=\"$" $1 "\"" }')"; fi
+  if \[ "$_otel_shell" = "dash" ]; then
+    \eval "$( { \printenv; \set; } | \grep '^OTEL_' | \cut -d= -f1 | \sort -u | \awk '{ print $1 "=\"$" $1 "\"" }' | _otel_line_join)" "\\$(_otel_escape_args "$@")"
+  else
+    "$@"
+  fi
 }
 
 otel_observe() {
@@ -219,15 +227,12 @@ otel_observe() {
   local command="${OTEL_SHELL_COMMANDLINE_OVERRIDE:-$*}"
   local command="${command#otel_observe }"
   local command="${command#_otel_observe }"
+  local command_signature="${OTEL_SHELL_COMMANDLINE_OVERRIDE_SIGNATURE:-$$}"
   local attributes="$OTEL_SHELL_SPAN_ATTRIBUTES_OVERRIDE"
-  if \[ -n "$OTEL_SHELL_ADDITIONAL_ARGUMENTS_POST_0" ]; then set -- "$@" "$(eval \\echo $OTEL_SHELL_ADDITIONAL_ARGUMENTS_POST_0)"; fi
-  if \[ -n "$OTEL_SHELL_ADDITIONAL_ARGUMENTS_POST_1" ]; then set -- "$@" "$(eval \\echo $OTEL_SHELL_ADDITIONAL_ARGUMENTS_POST_1)"; fi
   unset OTEL_SHELL_SPAN_NAME_OVERRIDE
   unset OTEL_SHELL_SPAN_KIND_OVERRIDE
   unset OTEL_SHELL_COMMANDLINE_OVERRIDE
   unset OTEL_SHELL_SPAN_ATTRIBUTES_OVERRIDE
-  unset OTEL_SHELL_ADDITIONAL_ARGUMENTS_POST_0
-  unset OTEL_SHELL_ADDITIONAL_ARGUMENTS_POST_1
   # create span, set initial attributes
   local span_id=$(otel_span_start $kind "$name")
   otel_span_attribute $span_id subprocess.executable.name=$(\echo "$command" | \cut -d' ' -f1 | \rev | \cut -d'/' -f1 | \rev)
@@ -236,16 +241,20 @@ otel_observe() {
   otel_span_attribute $span_id subprocess.command_args="$(\echo "$command" | \cut -sd' ' -f2-)"
   # run command
   otel_span_activate $span_id
+  if \[ -n "$OTEL_SHELL_ADDITIONAL_ARGUMENTS_POST_0" ]; then set -- "$@" "$(eval \\echo $OTEL_SHELL_ADDITIONAL_ARGUMENTS_POST_0)"; fi
+  if \[ -n "$OTEL_SHELL_ADDITIONAL_ARGUMENTS_POST_1" ]; then set -- "$@" "$(eval \\echo $OTEL_SHELL_ADDITIONAL_ARGUMENTS_POST_1)"; fi
+  unset OTEL_SHELL_ADDITIONAL_ARGUMENTS_POST_0
+  unset OTEL_SHELL_ADDITIONAL_ARGUMENTS_POST_1
   local exit_code=0
-  if \[ "$OTEL_SHELL_SUPPRESS_LOG_COLLECTION" != TRUE ]; then
+  if ! \[ -t 2 ] && \[ "$OTEL_SHELL_SUPPRESS_LOG_COLLECTION" != TRUE ]; then
     local traceparent=$OTEL_TRACEPARENT
     local stderr_pipe=$(\mktemp -u).opentelemetry_shell_$$.pipe
     \mkfifo $stderr_pipe
     ( (while IFS= read -r line; do otel_log_record $traceparent "$line"; \echo "$line" >&2; done < $stderr_pipe) & )
-    OTEL_SHELL_COMMANDLINE_OVERRIDE="$command" _otel_call "$@" 2> $stderr_pipe || local exit_code=$?
+    OTEL_SHELL_COMMANDLINE_OVERRIDE="$command" OTEL_SHELL_COMMANDLINE_OVERRIDE_SIGNATURE="$command_signature" _otel_call "$@" 2> $stderr_pipe || local exit_code=$?
     \rm $stderr_pipe
   else
-    OTEL_SHELL_COMMANDLINE_OVERRIDE="$command" _otel_call "$@" || local exit_code=$?
+    OTEL_SHELL_COMMANDLINE_OVERRIDE="$command" OTEL_SHELL_COMMANDLINE_OVERRIDE_SIGNATURE="$command_signature" _otel_call "$@" || local exit_code=$?
   fi
   otel_span_deactivate $span_id
   # set custom attributes, set final attributes, finish span
